Refactor model imports and remove deprecated base classes
