+++ contrawimae/models/__init__.py
@@ -2,7 +2,7 @@
 Model implementations for WiMAE and ContraWiMAE.
 """
 
-from .base import WiMAE
-from .contramae import ContraWiMAE
+from .wimae import WiMAE
+from .contrawimae import ContraWiMAE
 
 __all__ = ["WiMAE", "ContraWiMAE"]
